+++ R/map.R
@@ -37,7 +37,7 @@
 #' @param .at A character vector of names, positive numeric vector of
 #'   positions to include, or a negative numeric vector of positions to
 #'   exlude. Only those elements corresponding to `.at` will be modified.
-#' @param ... Additional arguments passed on to `.f`.
+#' @param ... Additional arguments passed on to the mapped function.
 #' @return All functions return a vector the same length as `.x`.
 #'
 #'   `map()` returns a list, `map_lgl()` a logical vector, `map_int()` an
@@ -91,6 +91,13 @@
 #' l2 %>% map(list("num", 3))
 #' l2 %>% map_int(list("num", 3), .default = NA)
 #'
+#'
+#' # Use a predicate function to decide whether to map a function:
+#' map_if(iris, is.factor, as.character)
+#'
+#' # Specify an alternative with the `.else` argument:
+#' map_if(iris, is.factor, as.character, .else = as.integer)
+#'
 #' # A more realistic example: split a data frame into pieces, fit a
 #' # model to each piece, summarise and extract R^2
 #' mtcars %>%
@@ -118,13 +125,20 @@ map <- function(.x, .f, ...) {
   .Call(map_impl, environment(), ".x", ".f", "list")
 }
 #' @rdname map
+#' @param .else A function applied to elements of `.x` for which `.p`
+#'   returns `FALSE`.
 #' @export
-map_if <- function(.x, .p, .f, ...) {
+map_if <- function(.x, .p, .f, ..., .else = NULL) {
   sel <- probe(.x, .p)
 
   out <- list_along(.x)
   out[sel]  <- map(.x[sel], .f, ...)
-  out[!sel] <- .x[!sel]
+
+  if (is_null(.else)) {
+    out[!sel] <- .x[!sel]
+  } else {
+    out[!sel]  <- map(.x[!sel], .else, ...)
+  }
 
   set_names(out, names(.x))
 }
@@ -139,7 +153,6 @@ map_at <- function(.x, .at, .f, ...) {
 
   set_names(out, names(.x))
 }
-
 
 #' @rdname map
 #' @export