+++ tests/testthat/test-map.R
@@ -104,3 +104,7 @@ test_that("primitive dispatch correctly", {
 test_that("map_if requires predicate functions", {
   expect_error(map_if(1:3, ~ NA, ~ "foo"), ", not a missing value")
 })
+
+test_that("`.else` maps false elements", {
+  expect_identical(map_if(-1:1, ~ .x > 0, paste, .else = ~ "bar", "suffix"), list("bar", "bar", "1 suffix"))
+})