+++ NEWS.md
@@ -64,6 +64,10 @@ reduce2_right(.x = letters[1:4], .y = paste2, .f = c("-", ".", "-")) # working
 
 ## Minor improvements and fixes
 
+* New `.else` argument for `map_if()` and `modify_if()`. They take an
+  alternative function that is mapped over elements of the input for
+  which the predicate function returns `FALSE` (#324).
+
 * Functions taking predicates (`map_if()`, `keep()`, `some()`,
   `every()`, `keep()`, etc) now fail with an informative message when
   the return value is not `TRUE` or `FALSE` (#470).